Item 1 - Nav Bar #2 - First Desktop Iteration Done!
Okay this looks good. animation is finally pretty polished. worked the UX a bit. seems good enough to have in mobile but i should get a proper mobile experience going. maybe like the one in the manga app.

but yeah the animation is smooth and precise and scroll navigation has checks and balances in place. think i'll get going on the home page next!