Item #...4? - Intro Page - Initial
Okay firebase is pretty integrated now with domain stuff all working good. intro page is good enough to call it an MVP i think! will deploy this.

need to make the intro way cooler tho. maybe like an AT A GLANCE with a bunch of fun widget or something.

Also i need to remember to look at this damn thing in 100% lmao, i keep developing in like 150% so things are mad small.