Item #4 - Intro Page - Dev #3
Okay!! got a significant amount of work done here, and finally have a decent direction for the intro at a glance module. guess i'm going with...shapes lol. actually, now that i'm looking at it it actually seems pretty metaphorical.

first we start with a basic shape, the triangle. this is my experience and what my career sits on. next comes the next shape with the tools in my toolbox. finally we have the most complex shape which outlines my proces, which in turn describes my values. not bad actually.

stoked to have the components all modular as well. much cleaner.

still desparately need to do the theme-ing. also the content for the shapes. lots of content actually. but feeling pretty proud of this chunk. lots to go but keep it going!!