Item #0 - Architecture - Dev #8
Lots of different stuff here, can hardly remember it all. some work on the aesthetic of the intro module, reorganization of the different modules in general.

this is labeled as architecture because i've decided to scrap the multi-route methodology to transition to just a single-page since that's all i'll really need. but the biggest change here is that each "route" will just be a module or group of modules, that will begin rendering when user has scrolled on them. the driver of this is the "page-scroll" functionality which FUNCTIONS JUST BY CSS. crazy. love that feature.

less javascript!! working pretty good. though the mobile view needs a lot of work and i think footer needs to be displaced. also the nav bar in general needs to be overhauled since it was written at the beginning, as well as is still using routing.

also i dont really have an organization for "items" in terms of my commits lol, seems pretty obscure at this point. OH and i upgraded all my packages. would like to use the latest React tech that i read an article about.