Item #0 - Architecture - Dev #9
Another big one!! rewired the nav like i was intending and got it connected with the observer watching the content component intializations, changed profession to my story and will get started on the timeline widget, got the scroll looking nice, and a bunch of other shit. is it....looking....good???? i think so.